Largely improve performance and do a major refactor
This commit drastically improves performance and stability of the animation process. The canvas is now a separate GUI component which internally holds an image which is used to save the needles' drawing progress.

A major refactoring of the entire project is performed: all logically separate parts of the program are now also physically separated, in terms of GUI components.